% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_plot.R
\name{abmMapPlot}
\alias{abmMapPlot}
\title{Plot simulation outputs}
\usage{
abmMapPlot(simResults, shelterLocations, avoidance = NULL, timeLimits = NA)
}
\arguments{
\item{simResults}{The resulting object created by \code{\link[=abm_simuate]{abm_simuate()}}.}

\item{shelterLocations}{The 2 column shelter location data.frame fed into the
\code{\link[=abm_simuate]{abm_simuate()}}.}

\item{avoidance}{The 2 column avoidance location data.frame fed into the
\code{\link[=abm_simuate]{abm_simuate()}} (optional).}

\item{timeLimits}{The start timestep of data to be plotted, can be helpful for
plotting a more manageable section of large simulation (optional).}
}
\value{
a plot
}
\description{
Plot simulation outputs
}
\seealso{
\code{\link[=abm_simuate]{abm_simuate()}} for the required input.
}
