% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_simulate.R
\name{abm_simulate}
\alias{abm_simulate}
\title{Agent based model: Simulate animal movement}
\usage{
abm_simulate(
  start,
  timesteps,
  des_options,
  options,
  shelterLocations,
  shelterSize,
  avoidPoints,
  destinationRange,
  destinationDirection,
  destinationTransformation,
  destinationModifier,
  avoidTransformation,
  avoidModifier,
  k_step,
  s_step,
  mu_angle,
  k_angle,
  rescale_step2cell = 1,
  behave_Tmat,
  rest_Cycle,
  additional_Cycles,
  shelteringMatrix,
  foragingMatrix,
  movementMatrix
)
}
\arguments{
\item{start}{A numeric vector of length 2, including the x and y coordinates
of the start location.}

\item{timesteps}{The number of time steps to be simulated.}

\item{des_options}{The number of dynamically chosen destinations presented to
the animal during the foraging behaviour state.}

\item{options}{The number of options the animal considers at each step.}

\item{shelterLocations}{A dataframe including the x and y coordinates for all
shelter sites, that act as points of attraction during the resting
behavioural state.}

\item{shelterSize}{A value describing the shelter site size. This value
dictates at what point the animal's movements will dramatically drop
simulating (near-)stationary behaviour.}

\item{avoidPoints}{A dataframe including the x and y coordinates that the
animal will avoid.}

\item{destinationRange}{A numeric vector of length two where the first value
provides the shape (k) parameter and the second value provides the scale
parameter (\eqn{\theta}) of the gamma distribution guiding the distance of
foraging destinations.}

\item{destinationDirection}{A numeric vector of length two where the first
value provides the mean (\eqn{\mu}) and the second value provides the
concentration (\eqn{\kappa}) of the Von Mises distribution guiding the
direction of foraging destinations.}

\item{destinationTransformation}{This parameter and the following three all
apply to the strength/pull/push the animal feels from a destination or
avoidance point. 0 - no transformation applied to the distance to
destination weighting, 1 - distance to destination weighing is
square-rooted, 2 - distance to destination weighting is squared}

\item{destinationModifier}{A coefficient to be applied to the distance to
destination weighting.}

\item{avoidTransformation}{Must be 0, 1, or 2: 0 = no transformation applied
to the distance to avoidance points weighting, 1 = distance to avoidance
points weighing is square-rooted, 2 = distance to avoidance points
weighting is squared.}

\item{avoidModifier}{A coefficient to be applied to the avoidance points
weighting.}

\item{k_step}{The shape parameters (k) for the gamma distribution describing
step length for each behavioural state. A vector of length 3.}

\item{s_step}{The scale parameters (\eqn{\theta}) for the gamma distribution
describing step length for each behavioural state. A vector of length 3.}

\item{mu_angle}{The means (\eqn{\mu}) for the Von Mises distribution used to
draw turn angles for each behavioural state. A vector of length 3.}

\item{k_angle}{The concentrations (\eqn{\kappa}) for the Von Mises
distribution used to draw turn angles for each behavioural state. A vector
of length 3.}

\item{rescale_step2cell}{A value that describes the cell size of the
environmental matrices relative to the units of the step lengths. Must be
greater than zero. E.g., step lengths described in metres, and using a
matrix where a cell is 1mx1m the rescale value would be 1. Whereas step
lengths in metres with a cell size of 10mx10m the rescale value would be
10. The step lengths returned are on the scale of the matrix and need to be
back transformed to match the input step length units for comparison.
Default is 1, step and matrix unit are the same.}

\item{behave_Tmat}{Base transition matrix for 3 behavioural states}

\item{rest_Cycle}{A vector length 4 for A, M, \eqn{\phi} and \eqn{\tau} to
define the resting/active cycle. Ideal for defining circadian rhythm.}

\item{additional_Cycles}{A (optional) data.frame 4 columns wide for A, M,
\eqn{\phi} and \eqn{\tau} to define any additional activity cycles. Ideal
for defining patterns that operate alongside circadian rhythm, e.g.,
seasonal shifts.}

\item{shelteringMatrix}{A matrix describing the sheltering quality of the
landscape. Higher values are the best shelter quality.}

\item{foragingMatrix}{A matrix describing the foraging quality of the
landscape. Higher values are the best foraging areas.}

\item{movementMatrix}{A matrix describing the movement ease of the landscape.
Higher values are the easiest to move through.}
}
\value{
A list with the following components: 1. "locations" The dataframe
describing all realised locations the animal occupied, where each row is
equal to a timestep. Columns include: timestep, the timestep as a integer;
x, the x coordinate of the animal; y, the y coordinate of the animal; sl,
the step length between locations used during the simulation; sl_rescale
the rescale factor required to return step lengths back to the input scale;
ta, turning angle between locations in degrees; behave, the behaviroual
mode the animal was in at a given timestep; chosen, the location chosen out
of the number of options available; destination_x and destination_y the
point the animal was attracted to at that time (note exploratory behaviour
is not subject attraction). 2. The "options" dataframe describing All the
options available to the animal over the entire simulation duration, where
each row is equal to an option repeated for each timestep. Columns include:
timestep, the timestep as a integer; oall_x, and oall_y show the x and y
coordinates of all the options available to an animal at a timestep;
oall_steplengths are the step lengths from the current location compared to
all the options. 3. The "inputs" return the arguments used to simulate the
movement, split into a list covering inputs_basic, inputs_destination,
inputs_movement, inputs_cycle, inputs_layerSeed. 4. "destinations" that
collects the choices available to the animal each time it shifts into a new
behavioural state. Columns include the timestep the choice was made at, and
the destinations x and y locations, and the behave the animal was moving
into. "others" captures all other outputs, mainly used internally for
debugging and checking.
}
\description{
An R function to arrange simulation parameters ready to be fed
to \emph{cpp_abm_simulate}, and also convert the list output of
\emph{cpp_abm_simulate} into a series of objects more easily used by downstream
functions.
}
\details{
The function automatically generates a list of seeds required for
the \emph{sampling_options} based upon any previous set R seeds set (\emph{e.g.},
using set.seed).
}
\seealso{
\code{\link[=vonmises]{vonmises()}} for further guidance on turn angle distribution
definition, \code{\link[=cycle_draw]{cycle_draw()}} for further guidance on activity cycle
definition. \code{\link[=cpp_abm_simulate]{cpp_abm_simulate()}} is the C++ function that is called.
}
