% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmAnimalMovement-package.R
\docType{package}
\name{abmAnimalMovement-package}
\alias{abmAnimalMovement}
\alias{abmAnimalMovement-package}
\title{abmAnimalMovement: An R Package for Simulating Animal Movement Using an Agent-based Model}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Benjamin Marshall \email{benjaminmichaelmarshall@gmail.com} (\href{https://orcid.org/0000-0001-9554-0605}{ORCID})

}
\keyword{internal}
