% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_basic_walk.R
\name{basic_walk}
\alias{basic_walk}
\title{R wrap for basic walk}
\usage{
basic_walk(start, steps, options, normmean, normsd, meanang, sdang)
}
\arguments{
\item{start}{The x and y coords of the starting location}

\item{steps}{The number of steps to be simulated}

\item{options}{The number of options to be considered at each step}

\item{normmean}{Parameter describing step length}

\item{normsd}{Parameter describing step angle}

\item{meanang}{Parameter describing angle}

\item{sdang}{Parameter describing angle variation}
}
\value{
Matrix of locations chosen
}
\description{
A function to run Cpp basic random walk where each step is chosen from a number of possible new locations.
}
