% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_abm_simulate}
\alias{cpp_abm_simulate}
\title{cpp_abm_simulate}
\arguments{
\item{startx}{The x coord start location}

\item{starty}{The y coord start location}

\item{steps}{The number of steps to be simulated}

\item{options}{The number of options to be considered at each step}

\item{normmean}{Parameter describing step length}

\item{normsd}{Parameter describing step angle}

\item{meanang}{Parameter describing angle}

\item{sdang}{Parameter describing angle variation}

\item{envMat1}{Environmental matrix 1}
}
\description{
Basic random walk
}
