% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_cycle_draw}
\alias{cpp_cycle_draw}
\title{cpp_cycle_draw}
\arguments{
\item{TIME}{A point in time (defined as a double) during the cycle to draw a value from.}

\item{A}{A double defining the amplitude of the cycle, calculated as the
difference from M (i.e., half of the overall variation in the cycle from
top to bottom).}

\item{M}{A double defining the cycle offset from 0 (Midline Statistic Of
Rhythm, a rhythm-adjusted mean).}

\item{PHI}{A double defining the offset of the cycle (\eqn{\phi}; i.e.,
acrophase). Must be kept proportional to \eqn{\tau} to avoid cycle drift.}

\item{TAU}{A double defining the cycle frequency (\eqn{\tau}; i.e., period),
where \eqn{\tau} is the difference between peaks.}
}
\value{
A double that is between M+A to M-A, draw from the defined cycle at TIME.
}
\description{
cpp_cycle_draw
}
\details{
\eqn{\phi} must remain proportional to \eqn{\tau} if the cycle is to remained
sync with period. If not proportional the cycle with drift.
}
\references{
Cornelissen, G. (2014). Cosinor-based rhythmometry. Theoretical
Biology and Medical Modelling. 11(16)
https://doi.org/10.1186/1742-4682-11-16
}
