% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_get_values}
\alias{cpp_get_values}
\title{cpp_get_values}
\value{
A vector of values
}
\description{
cpp_get_values
}
\details{
\eqn{\phi} must remain proportional to \eqn{\tau} if the cycle is to remained
sync with period. If not proportional the cycle with drift.
}
\references{
...
}
