% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_max/cpp_min}
\alias{cpp_max/cpp_min}
\title{Get maximum or minimum value}
\arguments{
\item{x}{A vector of doubles to the max/min found.}
}
\value{
The maximum or minimum value in the vector.
}
\description{
Get maximum or minimum value
}
