% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_vonmises}
\alias{cpp_vonmises}
\title{cpp_vonmises}
\usage{
cpp_vonmises(N, MU, KAPPA)
}
\arguments{
\item{N}{of values to draw.}

\item{MU}{The mean of the distribution (\eqn{\mu}).}

\item{KAPPA}{The concentration of the distribution (\eqn{\kappa}).}
}
\value{
A vector drawn from a von Mises distribution equal in length to N.
}
\description{
A C++ conversion of the \emph{rvm} function provided by the \href{https://CRAN.R-project.org/package=CircStats}{CircStats package}.
}
\references{
S-plus original by Ulric Lund and R port by Claudio Agostinelli (2018). CircStats: Circular Statistics, from Topics in Circular Statistics (2001). R package version 0.2-6. https://CRAN.R-project.org/package=CircStats
}
