% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_draw.R
\name{cycle_draw}
\alias{cycle_draw}
\title{Draw from cycle}
\usage{
cycle_draw(TIME, A, M, PHI, TAU)
}
\arguments{
\item{TIME}{A point in time during the cycle to draw a value from.}

\item{A}{A value defining the amplitude of the cycle, calculated as the
difference from M (i.e., half of the overall variation in the cycle from
top to bottom).}

\item{M}{A value defining the cycle offset from 0 (Midline Statistic Of
Rhythm, a rhythm-adjusted mean).}

\item{PHI}{A values defining the offset of the cycle (\eqn{\phi}; i.e.,
acrophase). Must be kept proportional to \eqn{\tau} to avoid cycle drift.}

\item{TAU}{A value defining the cycle frequency (\eqn{\tau}; i.e., period),
where \eqn{\tau} is the difference between peaks.}
}
\value{
A value between M+A and M-A draw from the defined cycle.
}
\description{
A function to define and draw from a cyclical distribution used
for daily of seasonal shifts.
}
\details{
The time intervals will match the scale the cycle was defined as.
For example, a \eqn{\tau} of 24 defines a cycle repeating every 24
instances, so a draw using TIME == 1 will extract a value at hour 1. For a
draw 2 and half hours in TIME == 2.5. Theta must be proportional to tau to
prevent the cycle from drifting. If kept proportional the cycle will remain
constrained to the \eqn{\tau} frequency defined. Implementation following
Cornelissen (2014).
}
\examples{

# an example of a cycle approximating nocturnal activity with a draw at hour 12.
cycle_draw(TIME = 12,
A = 1,
M = 0,
PHI = 4 / 24,
TAU = 24)

# an example of a cycle approximating crepuscular activity with a draw at hour 12.
cycle_draw(TIME = 12,
A = 1,
M = 0,
PHI = 24 / 12,
TAU = 12)

}
\references{
Cornelissen, G. (2014). Cosinor-based rhythmometry. Theoretical
Biology and Medical Modelling. 11(16)
https://doi.org/10.1186/1742-4682-11-16
}
