% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genLandscape.R
\name{genLandscape}
\alias{genLandscape}
\alias{genLandscape_noise}
\alias{genLandscape_gradient}
\alias{quick_plot_matrix}
\alias{genLandscape_quickTriple}
\title{Quick generate test landscapes}
\usage{
genLandscape_noise(row, col)

genLandscape_gradient(row, col)

quick_plot_matrix(mat)

genLandscape_quickTriple(row, col, seed)
}
\arguments{
\item{row, col}{number of rows and columns of the final matrix. Beware the
some functions will default with 100 then disaggregate to reach row and col
requirements. This is to save memory.}

\item{mat}{A matrix to plot using ggplot2}
}
\value{
a matrix usable as a environment layer / landscape
}
\description{
Quick generate test landscapes
}
