% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mult}
\alias{mult}
\title{Multiplies two doubles}
\usage{
mult(v1, v2)
}
\arguments{
\item{v1}{First value}

\item{v2}{Second value}
}
\value{
Product of v1 and v2
}
\description{
Multiplies two doubles
}
