% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/sample_test.R
\name{sample_test}
\alias{sample_test}
\title{sample_test}
\usage{
sample_test(vect)
}
\arguments{
\item{vect}{vector}
}
\value{
Matrix of locations chosen

chosen
}
\description{
sample_test

Sample test
}
