% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/vonmises.R
\name{vonmises}
\alias{vonmises}
\title{vonmises
https://statisticaloddsandends.wordpress.com/2020/02/25/what-is-the-von-mises-distribution/
https://gist.github.com/SiyuanQi/10ed29aa850e26e2e1251af04f5427ec
http://dirk.eddelbuettel.com/code/bh.html
https://garylarson.weebly.com/blog/category/rcpp}
\usage{
vonmises(N, kappa, mu, x)

vonmises(N, kappa, mu, x)
}
\arguments{
\item{N}{number to draw}
}
\value{
Vector pulled from vonmises

Vector pulled from vonmises
}
\description{
vonmises
https://statisticaloddsandends.wordpress.com/2020/02/25/what-is-the-von-mises-distribution/
https://gist.github.com/SiyuanQi/10ed29aa850e26e2e1251af04f5427ec
http://dirk.eddelbuettel.com/code/bh.html
https://garylarson.weebly.com/blog/category/rcpp

Vonmises
}
