% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_plot.R
\name{abm_plot}
\alias{abm_plot}
\alias{abmMapPlot}
\alias{abmCyclePlot}
\title{Plot simulation outputs}
\usage{
abmMapPlot(simResults, timeLimits = NA)

abmCyclePlot(simResults, timeLimits = NA)
}
\arguments{
\item{simResults}{The resulting object created by \code{\link[=abm_simuate]{abm_simuate()}}.}

\item{timeLimits}{The start timestep of data to be plotted, can be helpful for
plotting a more manageable section of large simulation (optional).}
}
\value{
Returns a ggplot2 plot
}
\description{
Plot simulation outputs
}
\seealso{
\code{\link[=abm_simuate]{abm_simuate()}} for the required input.
}
