% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_simulate.R
\name{abm_simulate}
\alias{abm_simulate}
\title{Agent based model: Simulate animal movement}
\usage{
abm_simulate(
  start,
  steps,
  options,
  k_step,
  s_step,
  mu_angle,
  k_angle,
  behave_Tmat,
  rest_Cycle,
  envMat1
)
}
\arguments{
\item{start}{A numeric vector of length 2, including the x and y coordinates
of the start location.}

\item{steps}{The number of time steps to be simulated, where each step is
equal to ------ ----.}

\item{options}{The number of options the animal considers at each step.}

\item{k_step}{The shape parameters (k) for the gamma distribution describing
step length for each behavioural state. A vector of length 3.}

\item{s_step}{The scale parameters (\eqn{\theta}) for the gamma distribution
describing step length for each behavioural state. A vector of length 3.}

\item{mu_angle}{The means (\eqn{\mu}) for the von Mises distribution used to
draw turn angles for each behavioural state. A vector of length 3.}

\item{k_angle}{The concentrations (\eqn{\kappa}) for the von Mises
distribution used to draw turn angles for each behavioural state. A vector
of length 3.}

\item{behave_Tmat}{Base transisition matrix for 3 behavioural states}

\item{rest_Cycle}{A vector length 4 for A, M, \eqn{\phi} and \eqn{\tau} to
define the resting/active cycle. Ideal for defining circadian rhythm.}

\item{envMat1}{TESTING ENVIRONMENTAL LAYER, A matrix.}
}
\value{
A list with the following components: 1. The location dataframe
describing all locations the animal occupied, where each row is equal to a
timestep. Columns include: - step, the step as a integer; - x, the x
coordinate of the animal; - y, the y coordinate of the animal; 2. The
options dataframe describing All the options available to the animal over
the entire simulation duration. Columns include: - a - b - c 3. ONWARDS...
TESTING OUTPUTS Columns include: - a - b
}
\description{
An R function to arrange simulation parameters ready to be fed
to \emph{cpp_abm_simulate}, and also convert the list output of
\emph{cpp_abm_simulate} into a series of objects more easily used by downstream
functions.
}
\details{
PROVIDE DETAILS THAT WOULD GUIDE INPUTS. The function automatically
generates a list of seeds required for the \emph{sampling_options} based upon
any previous set R seeds set (\emph{e.g.}, using set.seed).
}
