% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmAnimalMovement-package.R
\docType{package}
\name{abmAnimalMovement-package}
\alias{abmAnimalMovement}
\alias{abmAnimalMovement-package}
\title{abmAnimalMovement: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Benjamin Marshall \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
