% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_abm_simulate}
\alias{cpp_abm_simulate}
\title{cpp_abm_simulate}
\arguments{
\item{startx}{The x coord start location}

\item{starty}{The y coord start location}

\item{steps}{The number of steps to be simulated}

\item{options}{The number of options to be considered at each step}

\item{k_step}{Parameter describing step length}

\item{s_step}{Parameter describing step}

\item{mu_angle}{Parameter describing angle}

\item{k_angle}{Parameter describing angle variation}

\item{b0_Options}{Behave transitional probs for behave 0}

\item{b1_Options}{Behave transitional probs for behave 1}

\item{b2_Options}{Behave transitional probs for behave 2}

\item{rest_Cycle_A}{The amplitude of the resting/active cycle.}

\item{rest_Cycle_M}{The cycle offset from 0 (Midline Statistic Of Rhythm) of the resting/active cycle.}

\item{rest_Cycle_PHI}{The offset of the cycle (\eqn{\phi}; i.e., acrophase) of the resting/active cycle.}

\item{rest_Cycle_TAU}{Cycle frequency (\eqn{\tau}; i.e., period) of the resting/active cycle.}

\item{envMat1}{Environmental matrix 1}

\item{seeds}{A number of seeds the length of the number of steps, ideally generated by R handling function.}
}
\description{
The C++ function that runs the agent based animal movement model.
}
