% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_abm_simulate}
\alias{cpp_abm_simulate}
\title{cpp_abm_simulate}
\arguments{
\item{startx}{The x coordinate start location}

\item{starty}{The y coordinate start location}

\item{timesteps}{The number of timesteps to be simulated}

\item{ndes}{The number of possible destination considered for foraging.}

\item{nopt}{The number of options to be considered at each timestep}

\item{shelter_locs_x}{A vector of shelter locations x coordinates.}

\item{shelter_locs_y}{A vector of shelter locations y coordinates.}

\item{sSiteSize}{A double dictating at what range the animal's movements
will dramatically drop simulating (near-)stationary behaviour.}

\item{avoidPoints_x}{A vector of avoidance point x coordinates.}

\item{avoidPoints_y}{A vector of avoidance point y coordinates.}

\item{k_desRange}{Shape parameter describing the gamma distribution
destinations can be draw from.}

\item{s_desRange}{Scale parameter describing the gamma distribution
destinations can be draw from.}

\item{mu_desDir}{Mean of the turn angle destinations are drawn from.}

\item{k_desDir}{Concentration of the turn angles destinations are drawn
from.}

\item{destinationTrans}{0 - no transformation applied to the distance to
destination weighting, 1 - distance to destination weighing is
square-rooted, 2 - distance to destination weighting is squared}

\item{destinationMod}{A coefficient to be applied to the distance to
destination weighting.}

\item{avoidTrans}{0 - no transformation applied to the distance to avoidance
points weighting, 1 - distance to avoidance points weighing is
square-rooted, 2 - distance to avoidance points weighting is squared}

\item{avoidMod}{A coefficient to be applied to the avoidance points
weighting.}

\item{k_step}{Parameter describing step length}

\item{s_step}{Parameter describing step}

\item{mu_angle}{Parameter describing angle}

\item{k_angle}{Parameter describing angle variation}

\item{rescale}{A value that describes the cell size of the environmental
matrices relative to the units of the step lengths. Must be greater than
zero. The step lengths returned are on the scale of the matrix and need to
be back transformed to match the input step length units for comparison.
Default is 1, step and matrix unit are the same.}

\item{b0_Options}{Behave transitional probs for behave 0}

\item{b1_Options}{Behave transitional probs for behave 1}

\item{b2_Options}{Behave transitional probs for behave 2}

\item{rest_Cycle_A}{The amplitude of the resting/active cycle.}

\item{rest_Cycle_M}{The cycle offset from 0 (Midline Statistic Of Rhythm) of
the resting/active cycle.}

\item{rest_Cycle_PHI}{The offset of the cycle (\eqn{\phi}; i.e., acrophase)
of the resting/active cycle.}

\item{rest_Cycle_TAU}{Cycle frequency (\eqn{\tau}; i.e., period) of the
resting/active cycle.}

\item{addCycles}{A integer describing how many additional activity cycles to
include, can be zero.}

\item{add_Cycle_A}{The amplitude of the additional active cycles, should be
a vector of doubles.}

\item{add_Cycle_M}{The cycle offset from 0 (Midline Statistic Of Rhythm) of
the additional active cycles, should be a vector of doubles.}

\item{add_Cycle_PHI}{The offset of the cycle (\eqn{\phi}; i.e., acrophase)
of the additional active cycles, should be a vector of doubles.}

\item{add_Cycle_TAU}{Cycle frequency (\eqn{\tau}; i.e., period) of the
additional active cycles, should be a vector of doubles.}

\item{shelterMatrix}{The matrix describing shelter site quality.}

\item{forageMatrix}{The matrix describing foraging site quality.}

\item{moveMatrix}{The matrix describing movement ease.}
}
\value{
A list of simulated animal details to be passed and handled by R
function to make easier to use.
}
\description{
The C++ function that runs the agent based animal movement
model.
}
\details{
Requires cmath and headers for smaller C++ functions for draws.
}
\seealso{
\code{\link[=abm_simulate]{abm_simulate()}} is the R function wrap with more in depth
documentation.
}
