% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_get_values}
\alias{cpp_get_values}
\title{cpp_get_values}
\arguments{
\item{MATRIX}{A numeric matrix that the value will be extracted from.}

\item{XLOCS}{A vector of x locations.}

\item{YLOCS}{A vector of y locations.}
}
\value{
A vector of values equal to the length of XLOCS extracted from
MATRIX.
}
\description{
cpp_get_values
}
\details{
A simple C++ function to extract a value from a Rcpp::NumericMatrix
using two vectors describing location. Note that the counting begins for x
and y begins at zero.
}
