% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_sample_options}
\alias{cpp_sample_options}
\title{cpp_sample_options}
\arguments{
\item{W}{A vector of weights.}
}
\value{
An integer corresponding to the chosen weight in the initially
provided vector. \strong{NOTE: indexing begins at 0 matching C++ convention}.
}
\description{
The C++ sample function that takes a vector of weights and
randomly returns an integer of the choice using the Rcpp version of
sample().
}
\details{
Very simple wrapper to help handle std::vectors, and pre-normalise
the values provided to avoid issues with negative numbers.
}
