% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_vonmises}
\alias{cpp_vonmises}
\title{cpp_vonmises}
\usage{
cpp_vonmises(N, MU, KAPPA)
}
\arguments{
\item{N}{number to draw}
}
\value{
Vector pulled from vonmises
}
\description{
A C++ conversion of the rvm function provided by the CircStats package (https://cran.r-project.org/web/packages/CircStats/index.html)
}
\references{
S-plus original by Ulric Lund and R port by Claudio Agostinelli (2018). CircStats: Circular Statistics, from Topics in Circular Statistics (2001). R package version 0.2-6. https://CRAN.R-project.org/package=CircStats
}
