% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/sample_options.R
\name{sample_options}
\alias{sample_options}
\title{sample_options}
\usage{
sample_options(weights)
}
\arguments{
\item{weights}{probs}

\item{vect}{weights}
}
\value{
Matrix of locations chosen

chosen
}
\description{
sample_options

Sample options
}
