% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonmises.R
\name{vonmises}
\alias{vonmises}
\title{Draw from a von Mises distribution}
\usage{
vonmises(N, MU, KAPPA)
}
\arguments{
\item{N}{The number of values to draw.}

\item{MU}{The mean (\eqn{\mu}) direction of the distribution to be draw from.}

\item{KAPPA}{The concentration or shape parameter (\eqn{\kappa}). High value will provide a
tighter distribution around MU, whereas a lower value will widen the
distribution.}
}
\value{
Will return a vector the size of N draw from a Von Mises distribution
defined by \eqn{\mu} and \eqn{\kappa}. When \eqn{\mu} is zero the limits of the overall
distribution will be between -\eqn{\pi} and \eqn{\pi}. When considering \eqn{\kappa}, a higher
value will approximate a greater likely to draw values linked to straight
line movement.
}
\description{
This is the R function counterpart to the C++ function
\emph{cpp_vonmises}. The C++ function is based on the \emph{rvm} function from the
\href{https://cran.r-project.org/web/packages/CircStats/index.html}{CircStats package}.
}
\examples{
vonOut <- vonmises(N = 1000, MU = 0, KAPPA = 0.1)
hist(vonOut)

}
\references{
S-plus original by Ulric Lund and R port by Claudio Agostinelli
(2018). CircStats: Circular Statistics, from Topics in Circular Statistics
(2001). R package version 0.2-6. https://CRAN.R-project.org/package=CircStats
}
