% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonmises.R
\name{vonmises}
\alias{vonmises}
\title{Draw from a von Mises distribution}
\usage{
vonmises(N, MU, KAPPA)
}
\arguments{
\item{N}{The number of values to draw.}

\item{MU}{The mean direction of the distribution to be draw from.}

\item{KAPPA}{The concentration or shape parameter. High value will provide a
tighter distribution around MU, whereas a lower value will widen the
distribution.}
}
\value{
Will return a vector the size of N draw from a Von Mises distribution
defined by MU and KAPPA. When MU is zero the limits of the overall
distribution will be between -pi and pi. When considering KAPPA, a higher
value will approximate a greater likely to draw values linked to straight
line movement.
}
\description{
This is the R function counterpart to the c++ function
cpp_vonmises. The C++ function is based on the rvm function from the
CircStats package.
}
\examples{
vonOut <- vonmises(N = 1000, MU = 0, KAPPA = 0.1)
hist(vonOut)

}
\references{
S-plus original by Ulric Lund and R port by Claudio Agostinelli
(2018). CircStats: Circular Statistics, from Topics in Circular Statistics
(2001). R package version 0.2-6. https://CRAN.R-project.org/package=CircStats
}
