% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/vonmises.R
\name{vonmises}
\alias{vonmises}
\title{vonmises
Based on function fvm from teh CricStats package.}
\usage{
vonmises(N, MU, KAPPA)
}
\arguments{
\item{N}{number to draw}
}
\value{
Vector pulled from vonmises

Vector pulled from vonmises
}
\description{
vonmises
Based on function fvm from teh CricStats package.

Vonmises
}
