% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{walk_options_xy}
\alias{walk_options_xy}
\title{Basic random walk}
\arguments{
\item{startx}{The x coord start location}

\item{starty}{The y coord start location}

\item{steps}{The number of steps to be simulated}

\item{options}{The number of options to be considered at each step}

\item{normmean}{Parameter describing step length}

\item{normsd}{Parameter describing step angle}

\item{meanang}{Parameter describing angle}

\item{sdang}{Parameter describing angle variation}

\item{b0_Options}{Behave transitional probs for behave 0}

\item{b1_Options}{Behave transitional probs for behave 1}

\item{b2_Options}{Behave transitional probs for behave 2}

\item{envMat1}{Environmental matrix 1}
}
\value{
Matrix of locations chosen
}
\description{
Basic random walk
}
